requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dataclasses-json==0.5.14
Django==4.1.4
django-celery-results==2.5.1
django-environ==0.10.0
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
flower==2.0.0
frozenlist==1.4.0
greenlet==2.0.2
humanize==4.7.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.0
langchain==0.0.258
langsmith==0.0.20
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.24.4
openai==0.27.8
openapi-codec==1.3.2
openapi-schema-pydantic==1.2.4
packaging==23.1
Pillow==9.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.38
psycopg2==2.9.5
pydantic==1.10.12
PyJWT==2.6.0
pytz==2022.7
PyYAML==6.0.1
redis==4.5.5
requests==2.31.0
simplejson==3.19.1
six==1.16.0
SQLAlchemy==2.0.19
sqlparse==0.4.3
tenacity==8.2.2
tornado==6.3.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
django-timescaledb
